backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
google-genai==1.66.0
python-dotenv==1.0.0
pydantic>=2.9.0
Pillow